feat(settings/backup): better export indicator
Refs: #41
